qc_strands_project/requirements.txt
strands-agents
strands-agents-tools
strands-agents-builder

